+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.2.3] 2024-07-21
+
 ### Fixed
 
 - [Issue #3454144: Fix untranslated plant_type descriptions in plant and seed asset types](https://www.drupal.org/project/farm/issues/3454144)
@@ -784,7 +786,8 @@ moving forward.
     Drupal 7, which required a complete refactor of the codebase. By comparison,
     updating from Drupal 9 to 10 will simply involve updating deprecated code.
 
-[Unreleased]: https://github.com/farmOS/farmOS/compare/3.2.2...HEAD
+[Unreleased]: https://github.com/farmOS/farmOS/compare/3.2.3..HEAD
+[3.2.3]: https://github.com/farmOS/farmOS/releases/tag/3.2.3
 [3.2.2]: https://github.com/farmOS/farmOS/releases/tag/3.2.2
 [3.2.1]: https://github.com/farmOS/farmOS/releases/tag/3.2.1
 [3.2.0]: https://github.com/farmOS/farmOS/releases/tag/3.2.0